Remove unused ColorCompatibility properties
